+++ htfs/functions.go
@@ -5,13 +5,11 @@ import (
 	"crypto/sha256"
 	"fmt"
 	"io"
-	"math/rand"
 	"os"
 	"path/filepath"
 	"runtime"
 	"sort"
 	"sync"
-	"time"
 
 	"github.com/robocorp/rcc/anywork"
 	"github.com/robocorp/rcc/common"
@@ -219,54 +217,6 @@ func ScheduleLifters(library MutableLibrary, stats *stats) Treetop {
 	return scheduler
 }
 
-func TryRemove(context, target string) (err error) {
-	for delay := 0; delay < 5; delay += 1 {
-		time.Sleep(time.Duration(delay*100) * time.Millisecond)
-		err = os.Remove(target)
-		if err == nil {
-			return nil
-		}
-	}
-	return fmt.Errorf("Remove failure [%s, %s, %s], reason: %s", context, common.ControllerIdentity(), common.HolotreeSpace, err)
-}
-
-func TryRemoveAll(context, target string) (err error) {
-	for delay := 0; delay < 5; delay += 1 {
-		time.Sleep(time.Duration(delay*100) * time.Millisecond)
-		err = os.RemoveAll(target)
-		if err == nil {
-			return nil
-		}
-	}
-	return fmt.Errorf("RemoveAll failure [%s, %s, %s], reason: %s", context, common.ControllerIdentity(), common.HolotreeSpace, err)
-}
-
-func TryRename(context, source, target string) (err error) {
-	for delay := 0; delay < 5; delay += 1 {
-		time.Sleep(time.Duration(delay*100) * time.Millisecond)
-		err = os.Rename(source, target)
-		if err == nil {
-			return nil
-		}
-	}
-	common.Debug("Heads up: rename about to fail [%q -> %q], reason: %s", source, target, err)
-	origin := "source"
-	intermediate := fmt.Sprintf("%s.%d_%x", source, os.Getpid(), rand.Intn(4096))
-	err = os.Rename(source, intermediate)
-	if err == nil {
-		source = intermediate
-		origin = "target"
-	}
-	for delay := 0; delay < 5; delay += 1 {
-		time.Sleep(time.Duration(delay*100) * time.Millisecond)
-		err = os.Rename(source, target)
-		if err == nil {
-			return nil
-		}
-	}
-	return fmt.Errorf("Rename failure [%s, %s, %s, %s], reason: %s", context, common.ControllerIdentity(), common.HolotreeSpace, origin, err)
-}
-
 func LiftFile(sourcename, sinkname string) anywork.Work {
 	return func() {
 		source, err := os.Open(sourcename)
@@ -291,7 +241,7 @@ func LiftFile(sourcename, sinkname string) anywork.Work {
 
 		runtime.Gosched()
 
-		anywork.OnErrPanicCloseAll(TryRename("liftfile", partname, sinkname))
+		anywork.OnErrPanicCloseAll(pathlib.TryRename("liftfile", partname, sinkname))
 		pathlib.MakeSharedFile(sinkname)
 	}
 }
@@ -335,7 +285,7 @@ func DropFile(library Library, digest, sinkname string, details *File, rewrite [
 
 		anywork.OnErrPanicCloseAll(sink.Close())
 
-		anywork.OnErrPanicCloseAll(TryRename("dropfile", partname, sinkname))
+		anywork.OnErrPanicCloseAll(pathlib.TryRename("dropfile", partname, sinkname))
 
 		anywork.OnErrPanicCloseAll(os.Chmod(sinkname, details.Mode))
 		anywork.OnErrPanicCloseAll(os.Chtimes(sinkname, motherTime, motherTime))
@@ -344,13 +294,13 @@ func DropFile(library Library, digest, sinkname string, details *File, rewrite [
 
 func RemoveFile(filename string) anywork.Work {
 	return func() {
-		anywork.OnErrPanicCloseAll(TryRemove("file", filename))
+		anywork.OnErrPanicCloseAll(pathlib.TryRemove("file", filename))
 	}
 }
 
 func RemoveDirectory(dirname string) anywork.Work {
 	return func() {
-		anywork.OnErrPanicCloseAll(TryRemoveAll("directory", dirname))
+		anywork.OnErrPanicCloseAll(pathlib.TryRemoveAll("directory", dirname))
 	}
 }